+++ flash/__init__.py
@@ -24,8 +24,8 @@ if _TORCH_AVAILABLE:
     from flash.core.data.data_source import DataSource
     from flash.core.data.datasets import FlashDataset, FlashIterableDataset
     from flash.core.data.input_transform import InputTransform
-    from flash.core.data.io.output import Output, Serializer
-    from flash.core.data.process import Postprocess, Preprocess
+    from flash.core.data.io.output import Output
+    from flash.core.data.process import Postprocess, Preprocess, Serializer
     from flash.core.model import Task  # noqa: E402
     from flash.core.trainer import Trainer  # noqa: E402
 
@@ -47,9 +47,9 @@ if _TORCH_AVAILABLE:
         "FlashIterableDataset",
         "InputTransform",
         "Output",
-        "Serializer",
         "Postprocess",
         "Preprocess",
+        "Serializer",
         "Task",
         "Trainer",
     ]
